requirements.txt
asgiref==3.5.2
Django==4.1.3
django-countries==7.4.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
Pillow==9.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
